+++ include/cry/crc.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013-2014, Davide Galassi. All rights reserved.
+ * Copyright (c) 2013-2016, Davide Galassi. All rights reserved.
  *
  * This file is part of CRY software.
  *
@@ -29,6 +29,7 @@
 #define _CRY_CRC_H_
 
 #include <stddef.h>
+#include <stdint.h>
 
 /** CRC flags @{ */
 /** Complement result during finalization */
@@ -41,12 +42,23 @@
  * CRC-16 context structure
  */
 struct cry_crc16_ctx {
-    const unsigned short  *tab;
-    unsigned short        crc;
-    unsigned char         flags;
+    const uint16_t *tab;
+    uint16_t        crc;
+    unsigned char   flags;
 };
 
 typedef struct cry_crc16_ctx cry_crc16_ctx;
+
+/*
+ * CRC-32 context structure
+ */
+struct cry_crc32_ctx {
+    const uint32_t *tab;
+    uint32_t        crc;
+    unsigned char   flags;
+};
+
+typedef struct cry_crc32_ctx cry_crc32_ctx;
 
 #ifdef __cplusplus
 extern "C"{
@@ -60,8 +72,8 @@ extern "C"{
  * @param tab   precomputed lookup table.
  * @param flags algorithm behaviour flags.
  */
-void cry_crc16_init(struct cry_crc16_ctx *ctx, unsigned short start,
-                    const unsigned short *tab, unsigned char flags);
+void cry_crc16_init(struct cry_crc16_ctx *ctx, uint16_t start,
+                    const uint16_t *tab, unsigned char flags);
 
 /**
  * CRC update
@@ -79,7 +91,7 @@ void cry_crc16_update(struct cry_crc16_ctx *ctx,
  * @param ctx   CRC-16 context.
  * @return      CRC result.
  */
-unsigned short cry_crc16_final(struct cry_crc16_ctx *ctx);
+uint16_t cry_crc16_final(struct cry_crc16_ctx *ctx);
 
 /**
  * Initialize as a CRC-16-CCITT algorithm.
@@ -98,7 +110,7 @@ void cry_crc16_ccitt_init(struct cry_crc16_ctx *ctx);
  * @param n     Number of octets in the buffer.
  * @return      CRC value.
  */
-unsigned short cry_crc16_ccitt(const unsigned char *in, size_t n);
+uint16_t cry_crc16_ccitt(const unsigned char *in, size_t n);
 
 /**
  * Initialize as a CRC-16-IBM algorithm
@@ -119,7 +131,60 @@ void cry_crc16_ibm_init(struct cry_crc16_ctx *ctx);
  * @param n     Number of octets in the buffer.
  * @return      CRC value.
  */
-unsigned short cry_crc16_ibm(const unsigned char *in, size_t n);
+uint16_t cry_crc16_ibm(const unsigned char *in, size_t n);
+
+
+
+
+/**
+ * Initialize a generic CRC-32 algoritm context.
+ *
+ * @param ctx   CRC-32 context.
+ * @param start algorithm start value.
+ * @param tab   precomputed lookup table.
+ * @param flags algorithm behaviour flags.
+ */
+void cry_crc32_init(struct cry_crc32_ctx *ctx, uint32_t start,
+                    const uint32_t *tab, unsigned char flags);
+
+/**
+ * CRC update
+ *
+ * @param ctx   CRC-32 context.
+ * @param in    Buffer pointer.
+ * @param n     Number of octets in the buffer.
+ */
+void cry_crc32_update(struct cry_crc32_ctx *ctx,
+                      const unsigned char *in, size_t n);
+
+/**
+ * CRC finalize
+ *
+ * @param ctx   CRC-32 context.
+ * @return      CRC result.
+ */
+uint32_t cry_crc32_final(struct cry_crc32_ctx *ctx);
+
+/**
+ * Initialize as a CRC-32-ETH algorithm.
+ *
+ * Used by: Ethernet, ISO 3309, ANSI X3.66, FIPS PUB 71, ITU-T V.42, SATA,
+ *          MPEG-2, Gzip, PKZIP, POSIX cksum, PNG, ZMODEM.
+ * Polynomial:
+ *  X^32+X^26+X^23+X^22+X^16+X^12+X^11+X^10+X^8+X^7+X^5+X^4+X^2+X^1+X^0
+ *
+ * @param ctx   CRC-32 context.
+ */
+void cry_crc32_eth_init(struct cry_crc32_ctx *ctx);
+
+/**
+ * CRC-32-ETH all in one function.
+ *
+ * @param in    Buffer pointer.
+ * @param n     Number of octets in the buffer.
+ * @return      CRC value.
+ */
+uint32_t cry_crc32_eth(const unsigned char *in, size_t n);
 
 #ifdef __cplusplus
 }