+++ src/crc/crc16_ccitt.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013-2014, Davide Galassi. All rights reserved.
+ * Copyright (c) 2013-2016, Davide Galassi. All rights reserved.
  *
  * This file is part of CRY software.
  *
@@ -19,7 +19,7 @@
 
 #include "cry/crc.h"
 
-static const unsigned short table[] = {
+static const uint16_t table[] = {
     0x0000, 0x1189, 0x2312, 0x329b, 0x4624, 0x57ad, 0x6536, 0x74bf,
     0x8c48, 0x9dc1, 0xaf5a, 0xbed3, 0xca6c, 0xdbe5, 0xe97e, 0xf8f7,
     0x1081, 0x0108, 0x3393, 0x221a, 0x56a5, 0x472c, 0x75b7, 0x643e,
@@ -60,7 +60,7 @@ void cry_crc16_ccitt_init(struct cry_crc16_ctx *ctx)
                    CRY_CRC_FLAG_COMPLEMENT | CRY_CRC_FLAG_SWAP);
 }
 
-unsigned short cry_crc16_ccitt(const unsigned char *in, size_t n)
+uint16_t cry_crc16_ccitt(const unsigned char *in, size_t n)
 {
     struct cry_crc16_ctx ctx;